[not verified] move new block to v6/
