+++ projects/plugins/jetpack/extensions/blocks/videopress/v6/editor.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import registerJetpackBlock from '../../shared/register-jetpack-block';
+import registerJetpackBlock from '../shared/register-jetpack-block';
 import { name, settings } from '.';
 
 registerJetpackBlock( name, settings );
